+++ .storybook/preview.tsx
@@ -1,9 +1,17 @@
-import "../src/app/globals.css";
-
 import React from 'react';
+
 import type { Preview } from "@storybook/react";
 import { NotoSans } from "@/fontLoader.ts";
 import { css } from "~/styled-system/css";
+import { initialize, mswLoader } from 'msw-storybook-addon';
+
+import "../src/app/globals.css";
+
+initialize({
+  serviceWorker: {
+    url: "./mockServiceWorker.js"
+  }
+});
 
 const preview: Preview = {
   parameters: {
@@ -28,6 +36,7 @@ const preview: Preview = {
       </main>
     ),
   ],
+  loaders: [mswLoader]
 };
 
 export default preview;