+++ next-sitemap.config.js
@@ -1,6 +1,6 @@
 /** @type {import('next-sitemap').IConfig} */
 module.exports = {
-  siteUrl: process.env.SITE_URL || "https://kinguru.io",
+  siteUrl: process.env.SITE_URL || "https://eventify.today",
   output: "standalone",
   generateRobotsTxt: true,
 };