+++ panda.config.ts
@@ -1,9 +1,12 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import { defineConfig } from "@pandacss/dev";
+import { footerSlot } from "./src/components/Footer/Footer.recipe";
+import { headerSlot } from "./src/components/Header/Header.recipe";
 import { avatarRecipe } from "./src/components/uikit/Avatar/Avatar.recipe";
 import { buttonRecipe } from "./src/components/uikit/Button/Button.recipe";
 import { dropdownSlot } from "./src/components/uikit/Dropdown/Dropdown.recipe";
 import { inputRecipe } from "./src/components/uikit/Input/Input.recipe";
+import { selectSlot } from "./src/components/uikit/Select/Select.recipe";
 import { additionalGlobalCss } from "./src/theme/globalCss";
 import { layerStyles } from "./src/theme/layerStyles";
 import { customDividerPattern } from "./src/theme/patterns/customDivider";
@@ -33,6 +36,9 @@ export default defineConfig({
       },
       slotRecipes: {
         dropdown: dropdownSlot,
+        select: selectSlot,
+        footer: footerSlot,
+        header: headerSlot,
       },
       tokens: {
         gradients: {