+++ src/app/[locale]/layout.tsx
@@ -1,4 +1,6 @@
 import { NextIntlClientProvider, useMessages } from "next-intl";
+import { Footer } from "@/components/Footer";
+import { Header } from "@/components/Header";
 import { NotoSans } from "@/fontLoader.ts";
 import { css } from "~/styled-system/css";
 import { Box } from "~/styled-system/jsx";
@@ -25,30 +27,15 @@ export default function RootLayout({
     <html lang={locale} className={`${NotoSans.variable}`}>
       <NextIntlClientProvider locale={locale} messages={messages}>
         <body className={css({ fontFamily: "noto" })}>
+          <Header />
           <Box minHeight="100vh">
-            {/* Header placeholder START */}
-            <header
-              className={css({
-                zIndex: "1",
-                height: "85px",
-                bgColor: "neutral.5",
-                borderStyle: "solid",
-                borderColor: "neutral.3",
-                borderBottomWidth: "1px",
-                position: "sticky",
-                top: "0",
-              })}
-            />
-            {/* Header placeholder END */}
             <main
-              className={css({ height: "0", minHeight: "calc(100vh - 85px)" })}
+              className={css({ height: 0, minHeight: "calc(100vh - 85px)" })}
             >
               {children}
             </main>
           </Box>
-          {/* Footer placeholder START */}
-          <footer className={css({ height: "256px", bgColor: "neutral.1" })} />
-          {/* Footer placeholder END */}
+          <Footer />
         </body>
       </NextIntlClientProvider>
     </html>