+++ src/pages/[locale]/index.tsx
@@ -1,7 +1,7 @@
 import Head from "next/head";
 import { GetStaticPropsContext } from "next/types";
 import { useTranslations } from "next-intl";
-import { FooterSection } from "@/components/footer";
+import { FooterSection } from "@/components/footerOld";
 import {
   BestSpeakersSection,
   CompanyStatisticsSection,
@@ -27,7 +27,10 @@ export default function Home() {
           property="og:description"
           content={t("metadata.home.description")}
         />
-        <meta property="og:image" content="https://kinguru.io/img/brand.png" />
+        <meta
+          property="og:image"
+          content="https://eventify.today/img/brand.png"
+        />
       </Head>
       <Navbar />
       <HeroContent />