+++ src/pages/[locale]/price.tsx
@@ -1,7 +1,7 @@
 import Head from "next/head";
 import { GetStaticPropsContext } from "next/types";
 import { useTranslations } from "next-intl";
-import { FooterSection } from "@/components/footer";
+import { FooterSection } from "@/components/footerOld";
 import { HowItWorks } from "@/components/home";
 import { Navbar } from "@/components/navbar";
 import { OurOffers } from "@/components/pricing/ourOffers";
@@ -21,7 +21,10 @@ export default function Home() {
           property="og:description"
           content={t("metadata.price.description")}
         />
-        <meta property="og:image" content="https://kinguru.io/img/brand.png" />
+        <meta
+          property="og:image"
+          content="https://eventify.today/img/brand.png"
+        />
       </Head>
       <Navbar />
       <OurOffers />